Made photosets classes to better store date / id / title etc.
